# compiler: removes takeFreeRegisters fn